Creating nodes for multiple positions
